pyproject.toml
[tool.poetry]
name = "beancount-scripts"
version = "0.1.0"
description = ""
authors = ["Charlie Wren <23018417+charlietw@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "beancount_scripts"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.3"
pytest = "^7.2.2"
black = "^23.1.0"
beancount = "^2.3.5"
influxdb-client = "^1.36.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
